# Conflicts:
#	Presentation/Pages/UpdateArtwork.cshtml.cs
#	WebApiLayer/appsettings.json

+++ WebApiLayer/Controllers/ArtworkController.cs
@@ -1,6 +1,7 @@
 using BusinessLogicLayer.IService;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using ModelLayer.BussinessObject;
 using ModelLayer.DTOS.Request.Artwork;
 using ModelLayer.DTOS.Response;
@@ -12,10 +13,12 @@ namespace WebApiLayer.Controllers;
 public class ArtworkController : ControllerBase
 {
     private readonly IArtworkService _artworkService;
+    private readonly IHttpContextAccessor _contextAccessor;
 
-    public ArtworkController(IArtworkService artworkService)
+    public ArtworkController(IArtworkService artworkService, IHttpContextAccessor contextAccessor)
     {
         _artworkService = artworkService;
+        _contextAccessor = contextAccessor;
     }
 
     // GET: api/Artwork
@@ -31,12 +34,21 @@ public class ArtworkController : ControllerBase
     {
         return await _artworkService.GetArtworkByIdAsync(id);
     }
-
+    [Authorize]
+    [HttpGet]
+    public async Task<IEnumerable<ArtworkRespone>> GetOwnArtworks()
+    {
+        var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+        return await _artworkService.GetArtworkByArtistId(Guid.Parse(customer));
+    }
+    [Authorize]
     [HttpPost("create")]
     public async Task<IActionResult> CreateArtwork([FromForm] ArtworkCreation artworkCreation)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if(customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.AddArtworkAsync(artworkCreation);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -58,12 +70,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
-
+    [Authorize]
     [HttpPut("update")]
     public async Task<IActionResult> UpdateArtwork([FromForm]ArtworkUpdate artworkUpdate)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.UpdateArtworkAsync(artworkUpdate);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -84,12 +98,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
-
+    [Authorize]
     [HttpPost("remove/{id}")]
     public async Task<IActionResult> RemoveArtwork(Guid id)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.DeleteArtworkAsync(id);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -106,12 +122,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
-
+    [Authorize]
     [HttpPost("createTag")]
     public async Task<IActionResult> CreateTag4Artwork([FromForm]ArtworkTagAddition artworkTagAddition)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.AddTag4Artwork(artworkTagAddition);
             if(result is StatusCodeResult statusCodeResult)
             {
@@ -125,11 +143,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
+    [Authorize]
     [HttpPost("createCategory")]
     public async Task<IActionResult> CreateCategory4Artwork([FromForm] ArtworkCategoryAddition artworkCategoryAddition)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.AddCategory4Artwork(artworkCategoryAddition);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -143,11 +164,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
+    [Authorize]
     [HttpPut("updateTag")]
     public async Task<IActionResult> UpdateTag4Artwork([FromForm] ArtworkTagUpdate artworkTagUpdate)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.UpdateTag4Artwork(artworkTagUpdate);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -161,11 +185,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
+    [Authorize]
     [HttpPut("updateCategory")]
     public async Task<IActionResult> UpdateCategory4Artwork([FromForm] ArtworkCategoryUpdate artworkCategoryUpdate)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.UpdateCategory4Artwork(artworkCategoryUpdate);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -179,11 +206,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
+    [Authorize]
     [HttpPost("removeTag")]
     public async Task<IActionResult> RemoveTag4Artwork(Guid id)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.RemoveTag4Artwork(id);
             if(result is StatusCodeResult statusCodeResult)
             {
@@ -197,11 +227,14 @@ public class ArtworkController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
+    [Authorize]
     [HttpPost("removeCategory")]
     public async Task<IActionResult> RemoveCategory4Artwork(Guid id)
     {
         try
         {
+            var customer = _contextAccessor.HttpContext.User?.Claims?.FirstOrDefault(c => c.Type.Contains("Id")).Value;
+            if (customer == null) return StatusCode(StatusCodes.Status401Unauthorized);
             var result = await _artworkService.RemoveCategory4Artwork(id);
             if (result is StatusCodeResult statusCodeResult)
             {
@@ -227,7 +260,19 @@ public class ArtworkController : ControllerBase
     public async Task<List<ArtworkTag>> GetArtworkTagsByArtworkId(Guid id) => await _artworkService.GetArtworkTagByArtworkId(id);
     [HttpGet]
     public async Task<List<ArtworkTag>> GetArtworkTagsByTagId(Guid id) => await _artworkService.GetArtworkTagByTagId(id);
-    /*// GET: api/Artwork/5
+    [AllowAnonymous]
+    [HttpGet("resource")]
+    public async Task<ActionResult<List<Artwork>>> GetArtworkFromSearch([FromQuery] string search)
+    {
+        return await _artworkService.GetArtworkFromSearch(search);
+    }
+    [AllowAnonymous]
+	[HttpGet("{id}")]
+	public async Task<ActionResult<List<ArtworkRespone>>> GetArtworksByArtistId(Guid id)
+	{
+		return await _artworkService.GetArtworkByArtistId(id);
+	}
+	/*// GET: api/Artwork/5
     [HttpGet("{id}")]
     public async Task<ActionResult<Artwork>> GetArtwork(Guid id)
     {
